dsl_synctask: prevent signed overflow in MOS reservation
dsl_sync_task_common() converts blocks_modified to bytes with a
signed-int shift and stores the result in dsl_sync_task_t.dst_space.
dsl_sync_task_sync() then triples that value for the dittoed MOS
reservation. The multiplication exceeds INT_MAX when blocks_modified is
43,691, before the shift itself overflows at 131,072, causing spurious
failure with ENOSPC.

Recursive bookmark destruction can reach this boundary through normal
CLI use because it submits the Cartesian product of descendant datasets
and bookmark names as one sync task. At the first overflow boundary,
the undefined signed arithmetic can reject a valid destroy with
ENOSPC.

Make dst_space uint64_t and cast blocks_modified before shifting it.
This keeps dst_space in bytes and makes the existing * 3 reservation
formula operate in uint64_t. The dsl_sync_task() parameter and exported
function signatures remain unchanged; no ioctl, nvlist, or on-disk
format changes.

Add a zfs_destroy_bookmarks regression that submits 43,692 recursive
targets using missing bookmark names. Missing names are successful
no-ops but exercise the same batched kernel path, covering the boundary
without creating or destroying tens of thousands of bookmarks.

+++ include/sys/dsl_synctask.h
@@ -91,7 +91,7 @@ typedef struct dsl_sync_task {
 	txg_node_t dst_node;
 	struct dsl_pool *dst_pool;
 	uint64_t dst_txg;
-	int dst_space;
+	uint64_t dst_space;
 	zfs_space_check_t dst_space_check;
 	dsl_checkfunc_t *dst_checkfunc;
 	dsl_syncfunc_t *dst_syncfunc;
